Fix Makefile to use new immac for initial make
When doing git pull on a computer, which had an older version of the
pre-compiler immac installed, which had no -Y option, the initial
make, which involves pre-compiling comp.y to remove BOOT_COMPILE code,
failed (unknown option -Y) because make used the previously installed
version of immac.

The failure was exacerbated by a bug in immac, which caused a very
long loop when immac was called with no option -m, -M or -Y with a
file which did not have the .ica extension, but which contained .ica
keywords FOR IF ELSE etc (comp.y contains most of these).  This caused
.ica file keyword processing, which is not appropriate. This caused
the problem.

Makefile.in  use newest ./immac -Y to pre-compile comp.y
immac        force -m option for non .ica files - not just .ic files

    Id: src/Makefile.in 1.71
    Id: src/immac 1.26