Implement secondary I/O Equivalences in iCserver
In very rare cases equivalencing in iCserver did not work
correctly. This happened for two cases:

1)  when equivalencing several inputs for the same instance of an app.
2)  when equivalencing an output of an app to an input of the same app.

These cases are rather obscure but they may be of interest with real
I/O applications. In the previous implementation the Sender channel
(called the primary channel) could only be linked by an equivalence to
one Input Receiver on that channel on any one client app. To allow more
equivalenced Input Receivers to be linked to the same instance, one
or more secondary channels are defined which form a set of channels,
to which the Sender sends its output when the Sender receives a new
message on the primary channel.

Another change is to iCbox, which can now also start a client in a
Bernstein chain with the -R option. This makes Bernstein chaining
uniform for all clients and the server iCserver.

iCbox     implements -R Bernstein chaining
iClive    changes in comments only
iCserver  implement secondary I/O Equivalences
iC.odt    Note on language changes for iC version 3
iC.pdf    same

    Id: src/iCbox 1.62
    Id: src/iClive 1.69
    Id: src/iCserver 1.54
    Id: doc/iC.odt 1.55
    Id: doc/iC.pdf 1.55