Implement debug Watch and Ignore points for iClive
The immediate C Integrated Development Environment (IDE) program
iClive had facilities for editing and compiling code and inspecting
the state and values of iC variables while a program is running for
a long time. The usual debugging facilities of break points and the
ability to single step a program has been missing so far.

Break points in a data driven environment do not make any sense,
because there is no linear instruction flow of the iC code. Execution
is driven by changes in value of immediate variables. This means a
single execution step in a data driven environment is a value change
in an iC variable, which so far led to a message being sent to iClive
to update the live display of that variable.

What was required is the implementation of 'Watch' points, which
are also available in modern conventional debuggers. A Watch point
nominates a variable (not an instruction) and when that variable
changes the execution stops. As an option, a Watch point can be
adorned with a condition - in that case execution only stops when
the condition is true.

To implement this scheme extra care was taken not to slow down iC
run-time execution. Only one extra if statement has been inserted
in the run-time code to test if debugging is on and only two extra
instructions to recognise a Watch point. The bulk of the Watch point
code is in iClive and that only executes when a break message has
been received from the run-time code. At that point the run-time is
frozen and time is no longer critical.

Closely related to Watch points is single stepping. Four levels of
single stepping have been implemented.

1)  'Step' when the source text is the iC source text. In this case
    a change in every iC variable in the source causes a break.
    Changes in intermediate variables generated by the compiler are
    ignored.

2)  'Step' when the source text is the iC listing text, which shows
    every intermediate variable generated by the compiler. A change
    in all these variables causes a break leading to more fine grained
    inspection of the progress of execution.

3)  'Next' causes a break for a change in every output variable only.

4)  'Continue' causes a break for a change in every Watch point only.

When a break occurs, the state of the break variable is indicated
by its colour as usual, but additionally every instance of the break
variable is made prominent by surrounding it with a graphical border
which makes it look raised. Additionally the variable text is made
bold and the variable name is shown in the Search Entry window which
turns from light green (indicating debugging is active) to pink at a
break. If no instance of the break variable is visible in the current
text window, a search is started to move the text in the window to
show it.

Another useful feature is TextDebug output, which can be switched
on and off while debugging. This is the equivalent of machine code
in a conventional instruction flow computer. This feature has been
part if the iC system since early in its development to visualize
what is happining in detail in the run-time system. But without
breaks it was always hard to use because of the continuous changes
in a running system. With breaks the recent history of intermediate
variable changes before the break can be analyzed at leisure. The
following is an example of an intermediate variable b0_4 setting a
D flip flop b0_1, which switches a gate b0 from 0 to 1 (+1 to -1)
at which point the break occured.

b0_4:   -1      b0_1 +1 D=>> -1
b0_1:   -1      b0 +1 +=>> -1

A running program can be interrupted at any time when debugging is
switched on by pressing 'Step' or 'Next'.

Another feature of the new iClive debugger are 'Ignore' points, which
have no equivalent in conventional debuggers. A variable marked as an
Ignore point will not break when single stepping. This can be used for
marking sections of the iC program which are not part of the sub-set
of code currently being investigated and whose functionality should
keep running and changes in that part of the code should not interrupt
the sub-set being investigated. A good example is code controlled by
timer inputs. For this reason the timer inputs TX0.x themselves are
marked as permanent Ignore points. A break can only be generated by
timer inputs by setting a Watch point on them.

README      iC rev icc_3.2
comp.h      use global variables sTable and sEnd for iC_buildNet() and iC_icc() to define common Symbol Table
iClive      Holding the Live button down opens a menu to start and stop Debug mode and control options
	    Provide facilities to enter and change Watch and Ignore points
	    Interpret BREAK messages from an app for Watch breaks and single step breaks
            Introduce a Word mode search and make it the default
iCpiFace.c  Make input, extra input and timer event parameters of iC_wait_for_next_event() to make them optional
iCpiPWM.c   ditto
iCserver    Allow messages to be received from any connection. This allows apps to re-transmit received messages for debugging
icbegin.c   Copyright 2017
icc.c       use global variables sTable and sEnd for iC_buildNet() and iC_icc() to define common Symbol Table
icc.h       ditto
icend.c     Copyright 2017
icr.c       use global variables sTable and sEnd for iC_buildNet() and iC_icc() to define common Symbol Table
ict.c       use global variables sTable and sEnd for iC_buildNet() and iC_icc() to define common Symbol Table
            Add code to insert and use bits in Symbol Table to generate breaks to implement Watch and Ignore points
link.c      Send iC_liveData() after result of every change in TextDebug output has been finalised
load.c      use global variables sTable and sEnd for iC_buildNet() and iC_icc() to define common Symbol Table
outp.c      ditto
p_channel.c Copyright 2017
rsff.c      Send iC_liveData() after result of every change in TextDebug output has been finalised
            Fix error - Send iC_liveData() for master action of RISE, CHANGE bit and CHANGE int
scan.c      Send iC_liveData() after result of every change in TextDebug output has been finalised
symb.c      Copyright 2017
tcpc.c      Make input, extra input and timer event parameters of iC_wait_for_next_event() to make them optional
tcpc.h      ditto

    Id: src/README 1.31
    Id: src/comp.h 1.74
    Id: src/iClive 1.70
    Id: src/iCpiFace.c 1.9
    Id: src/iCpiPWM.c 1.5
    Id: src/iCserver 1.55
    Id: src/icbegin.c 1.8
    Id: src/icc.c 1.81
    Id: src/icc.h 1.83
    Id: src/icend.c 1.5 <- icend.c,v
    Id: src/icr.c 1.46
    Id: src/ict.c 1.72
    Id: src/link.c 1.34 <- link.c,v
    Id: src/load.c 1.68
    Id: src/outp.c 1.108
    Id: src/p_channel.c 1.3 <- p_channel.c,v
    Id: src/rsff.c 1.62
    Id: src/scan.c 1.44
    Id: src/symb.c 1.25
    Id: src/tcpc.c 1.28 <- tcpc.c,v
    Id: src/tcpc.h 1.22 <- tcpc.h,v