Modify gaIdke-pre-commit to exclude deleted files
Remove all remaining sources generated by bison or yacc and flex
to compile RASPBERRYPI Raspbian and other platforms correctly.
In the previous commit, gaIdke-pre-commit saw the change from the
current file to nothing for the files to be deleted and marked them
with an updated revision and then added them back to the index.

../.git_filters/gaIdke-pre-commit: exclude deleted files from revision update.

	deleted:    comp.tab.c
	deleted:    cons.tab.c
	deleted:    gram.tab.c
	deleted:    lex.yy.c

    Id: .git_filters/gaIdke-pre-commit 1.15