Retain source $Id's for files generated by bison and flex
Previously the $Id <path> was changed to the name of the generated
sources fom .y bison and .l flex sources, because the generated sources
were also kept under GIT version control.  Since the previous commit
this is no longer the case. Therefore it is appropriate that the name
of the orginal .y and .l source file is passed down to the generated
file and from there to the object and execuatable files.

Makefile.in: do not change $Id <path> to generated source when calling bison or flex

    Id: src/Makefile.in 1.74