Modify gaIdke-post-commit to smudge correctly after git pull
In GIT versions prior to 2.6.6 'git checkout HEAD' would execute
unconditionally and smudge the file. This no longer happens if
the file to be checked out is identical to the GIT archive file
and has a time stamp of the same second as the GIT archive file.
For this reason the .gaidke backup is moved back to make it
different to the committed file, which also removes it, because
the checked out file overwrites it with the committed file.

If called via 'post-checkout', there is no backup file $path.gaidke.
To smudge correctly for GIT >= 2.6.6, $path in working directory
must be removed.

.git_filters/gaIdke-activate:    change error message from Oops to Error:
.git_filters/gaIdke-post-commit: remove $path from WD if called via 'post-checkout'
.git_filters/gaIdke-pre-commit:  change error message from Oops to Error:
.git_filters/gaIdke-reset:       change error message from Oops to Error:

    Id: .git_filters/gaIdke-activate 1.5
    Id: .git_filters/gaIdke-post-commit 1.10
    Id: .git_filters/gaIdke-pre-commit 1.16
    Id: .git_filters/gaIdke-reset 1.2